Separate from PenicillinAction.kt
