+++ src/main/kotlin/jp/nephy/penicillin/core/request/action/JsonRequest.kt
@@ -1,0 +1,8 @@
+package jp.nephy.penicillin.core.request.action
+
+import jp.nephy.penicillin.models.PenicillinModel
+import kotlin.reflect.KClass
+
+interface JsonRequest<M: PenicillinModel> {
+    val model: KClass<M>
+}
